requirements.txt
streamlit
langchain_huggingface
langchain_community
torch
langchain
python-dotenv
asyncio
langchain_groq
Watchdog
numpy==1.26.2
playwright
html2text
faiss-cpu